projects/Jules Trading Platform/Master/requirements.txt
python-binance
websockets
python-dotenv
pandas
numpy
requests
vaderSentiment
pandas-ta
Flask
